# petstore/static/swagger.json
{
    "schemes": [
      "https",
      "http"
    ],
    "swagger": "2.0",
    "info": {
      "description": "This is a sample server Petstore server. ",
      "title": "Swagger Petstore",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0.7"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
    "paths": {
      "/pet": {
        "put": {
          "security": [
            {
              "petstore_auth": [
                "write:pets",
                "read:pets"
              ]
            }
          ],
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "pet"
          ],
          "summary": "Update an existing pet",
          "operationId": "updatePet",
          "parameters": [
            {
              "description": "Pet object that needs to be added to the store",
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Pet"
              }
            }
          ]
        },
        "post": {
          "security": [
            {
              "petstore_auth": [
                "write:pets",
                "read:pets"
              ]
            }
          ],
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "pet"
          ],
          "summary": "Add a new pet to the store",
          "operationId": "addPet",
          "parameters": [
            {
              "description": "Pet object that needs to be added to the store",
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Pet"
              }
            }
          ]
        }
      },
      "/pet/findByStatus": {
        "get": {
          "security": [
            {
              "petstore_auth": [
                "write:pets",
                "read:pets"
              ]
            }
          ],
          "description": "Multiple status values can be provided with comma separated strings",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "pet"
          ],
          "summary": "Finds Pets by status",
          "operationId": "findPetsByStatus",
          "parameters": [
            {
              "type": "array",
              "items": {
                "enum": [
                  "available",
                  "pending",
                  "sold"
                ],
                "type": "string",
                "default": "available"
              },
              "collectionFormat": "multi",
              "description": "Status values that need to be considered for filter",
              "name": "status",
              "in": "query",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Pet"
                }
              }
            }
          }
        }
      },
...
      "/pet/{petId}": {
        "get": {
          "security": [
            {
              "api_key": []
            }
          ],
          "description": "Returns a single pet",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "pet"
          ],
          "summary": "Find pet by ID",
          "operationId": "getPetById",
          "parameters": [
            {
              "type": "integer",
              "format": "int64",
              "description": "ID of pet to return",
              "name": "petId",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Pet"
              }
            }
          }
        },
        "post": {
          "security": [
            {
              "petstore_auth": [
                "write:pets",
                "read:pets"
              ]
            }
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "pet"
          ],
          "summary": "Updates a pet in the store with form data",
          "operationId": "updatePetWithForm",
          "parameters": [
            {
              "type": "integer",
              "format": "int64",
              "description": "ID of pet that needs to be updated",
              "name": "petId",
              "in": "path",
              "required": true
            },
            {
              "type": "string",
              "description": "Updated name of the pet",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "Updated status of the pet",
              "name": "status",
              "in": "formData"
            }
          ]
        },
        "delete": {
          "security": [
            {
              "petstore_auth": [
                "write:pets",
                "read:pets"
              ]
            }
          ],
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "pet"
          ],
          "summary": "Deletes a pet",
          "operationId": "deletePet",
          "parameters": [
            {
              "type": "string",
              "name": "api_key",
              "in": "header"
            },
            {
              "type": "integer",
              "format": "int64",
              "description": "Pet id to delete",
              "name": "petId",
              "in": "path",
              "required": true
            }
          ]
        }
      },
      "/store/inventory": {
        "get": {
          "security": [
            {
              "api_key": []
            }
          ],
          "description": "Returns a map of status codes to quantities",
          "produces": [
            "application/json"
          ],
          "tags": [
            "store"
          ],
          "summary": "Returns pet inventories by status",
          "operationId": "getInventory",
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
...
      "/store/order/{orderId}": {
        "get": {
          "description": "For valid response try integer IDs with value \u003e= 1 and \u003c= 10. Other values will generated exceptions",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "store"
          ],
          "summary": "Find purchase order by ID",
          "operationId": "getOrderById",
          "parameters": [
            {
              "maximum": 10,
              "minimum": 1,
              "type": "integer",
              "format": "int64",
              "description": "ID of pet that needs to be fetched",
              "name": "orderId",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Order"
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            }
          }
        },
        "delete": {
          "description": "For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "store"
          ],
          "summary": "Delete purchase order by ID",
          "operationId": "deleteOrder",
          "parameters": [
            {
              "minimum": 1,
              "type": "integer",
              "format": "int64",
              "description": "ID of the order that needs to be deleted",
              "name": "orderId",
              "in": "path",
              "required": true
            }
          ]
        }
      },
...
    },
    "definitions": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Category"
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "complete": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "petId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "shipDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "description": "Order Status",
            "type": "string",
            "enum": [
              "placed",
              "approved",
              "delivered"
            ]
          }
        },
        "xml": {
          "name": "Order"
        }
      },
      "Pet": {
        "type": "object",
        "required": [
          "name",
          "photoUrls"
        ],
        "properties": {
          "category": {
            "$ref": "#/definitions/Category"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "example": "doggie"
          },
          "photoUrls": {
            "type": "array",
            "items": {
              "type": "string",
              "xml": {
                "name": "photoUrl"
              }
            },
            "xml": {
              "wrapped": true
            }
          },
          "status": {
            "description": "pet status in the store",
            "type": "string",
            "enum": [
              "available",
              "pending",
              "sold"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Tag",
              "xml": {
                "name": "tag"
              }
            },
            "xml": {
              "wrapped": true
            }
          }
        },
        "xml": {
          "name": "Pet"
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Tag"
        }
      },
...
    },
    "securityDefinitions": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      },
      "petstore_auth": {
        "type": "oauth2",
        "flow": "implicit",
        "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
        "scopes": {
          "read:pets": "read your pets",
          "write:pets": "modify pets in your account"
        }
      }
    },
...
  }